Fixed an issue with the saving of the log. Changed to a simpler connect to room function. Made the synchronizing of the puzzle only include the data that has changed. Made it not possible to attach jigsaw pieces if either are still grabbed. Added a max grabber count of 1 to jigsaw pieces when they're spawned.

+++ Assets/Core/Scripts/InfoLogger.cs
@@ -38,7 +38,7 @@ public class InfoLogger : MonoBehaviour
             myLog += mylog;
         }
 
-        if (!string.IsNullOrEmpty(logPath))
+        if (!string.IsNullOrEmpty(logPath) && Directory.Exists(logPath))
         {
             if (!File.Exists(logPath))
                 File.Create(logPath);
